Move GameRendering to component + breakdown
