+++ frontend/src/components/Game/Game.js
@@ -1,0 +1,3 @@
+export default function Game() {
+    return <div>GAME WILL BE HERE</div>
+}
