+++ frontend/src/components/Highscore.js
@@ -1,0 +1,8 @@
+useEffect(() => {
+    async function loadHighscores() {
+      const res = await fetch('./highscores');
+      const payload = await res.json();
+      console.log(payload.data);
+    }
+    loadHighscores();
+  }, []);
